+++ CMake/Tests.cmake
@@ -41,6 +41,7 @@ set(tests
   stores_test
   tile_properties_test
   timedemo_test
+  tile_migration_sync_test
   townerdat_test
   writehero_test
   vendor_test